% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_node.R
\name{AgentMapNode}
\alias{AgentMapNode}
\title{Agent Map Node R6 Class}
\value{
An `AgentMapNode` object.
}
\description{
A node that maps over a list in the state and performs an operation.
}
\section{Super class}{
\code{\link[HydraR:AgentNode]{HydraR::AgentNode}} -> \code{AgentMapNode}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{map_key}}{String. Key in state to map over.}

\item{\code{logic_fn}}{Function(item, state) -> List(status, output).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AgentMapNode-new}{\code{AgentMapNode$new()}}
\item \href{#method-AgentMapNode-run}{\code{AgentMapNode$run()}}
\item \href{#method-AgentMapNode-clone}{\code{AgentMapNode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentMapNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-AgentMapNode-new}{}}}
\subsection{Method \code{new()}}{
Initialize AgentMapNode
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentMapNode$new(id, map_key, logic_fn, label = NULL, params = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Unique identifier.}

\item{\code{map_key}}{String identifier for state retrieval.}

\item{\code{logic_fn}}{Mapping function.}

\item{\code{label}}{Optional label.}

\item{\code{params}}{Optional parameters.
Run the Map Node}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentMapNode-run"></a>}}
\if{latex}{\out{\hypertarget{method-AgentMapNode-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentMapNode$run(state, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{state}}{AgentState object.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with status, output (list of results).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AgentMapNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AgentMapNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AgentMapNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
