% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_edge_synchronization}
\alias{check_edge_synchronization}
\title{Check Edge Synchronization}
\usage{
check_edge_synchronization(dag)
}
\arguments{
\item{dag}{AgentDAG.}
}
\value{
List(errors).
}
\description{
Check Edge Synchronization
}
\keyword{internal}
