% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_resource_linking}
\alias{check_resource_linking}
\title{Check Resource Linking (Roles & Logic)}
\usage{
check_resource_linking(dag)
}
\arguments{
\item{dag}{AgentDAG.}
}
\value{
List(errors, warnings).
}
\description{
Check Resource Linking (Roles & Logic)
}
\keyword{internal}
