% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{.build_llm_node}
\alias{.build_llm_node}
\title{Internal helper: Build an LLM Node}
\usage{
.build_llm_node(id, label, params, drv_reg)
}
\description{
Internal helper: Build an LLM Node
}
\keyword{internal}
