% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{.build_logic_node}
\alias{.build_logic_node}
\title{Internal helper: Build a Logic Node}
\usage{
.build_logic_node(id, label, params)
}
\description{
Internal helper: Build a Logic Node
}
\keyword{internal}
