% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{.build_observer_node}
\alias{.build_observer_node}
\title{Internal helper: Build an Observer Node}
\usage{
.build_observer_node(id, label, params)
}
\description{
Internal helper: Build an Observer Node
}
\keyword{internal}
