% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{.build_router_node}
\alias{.build_router_node}
\title{Internal helper: Build a Router Node}
\usage{
.build_router_node(id, label, params)
}
\description{
Internal helper: Build a Router Node
}
\keyword{internal}
