% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\docType{data}
\name{.hydra_registry}
\alias{.hydra_registry}
\title{HydraR Logic Registry}
\format{
An object of class \code{environment} of length 2.
}
\usage{
.hydra_registry
}
\description{
An internal environment mapping names to functions for reducers and schemas.
This allows saving AgentState as JSON by only storing function names.
}
\keyword{internal}
