% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.lint_single_logic_block}
\alias{.lint_single_logic_block}
\title{Internal helper: Lint a single logic block}
\usage{
.lint_single_logic_block(code, name)
}
\description{
Internal helper: Lint a single logic block
}
\keyword{internal}
