% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{.resolve_file_pattern}
\alias{.resolve_file_pattern}
\title{Internal handler: Resolve File Logic Pattern}
\usage{
.resolve_file_pattern(v_trim, base_dir)
}
\description{
Internal handler: Resolve File Logic Pattern
}
\keyword{internal}
