% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{.resolve_registered_pattern}
\alias{.resolve_registered_pattern}
\title{Internal handler: Resolve Registry/Global Pattern}
\usage{
.resolve_registered_pattern(v_trim)
}
\description{
Internal handler: Resolve Registry/Global Pattern
}
\keyword{internal}
