% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{format_toolset}
\alias{format_toolset}
\title{Format Toolset for Prompt}
\usage{
format_toolset(tools)
}
\arguments{
\item{tools}{List of AgentTool objects.}
}
\value{
A formatted string containing all tool descriptions.
}
\description{
Format Toolset for Prompt
}
\examples{
\dontrun{
# Format a collection of tools for an agent
tools <- list(
  search = AgentTool$new("google_search", "Search the web"),
  run_r = AgentTool$new("r_exec", "Execute R code locally")
)

prompt_appendix <- format_toolset(tools)
cat(prompt_appendix)
}
}
