% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_registry.R
\name{get_default_driver}
\alias{get_default_driver}
\title{Get the Default Agent Driver}
\usage{
get_default_driver()
}
\value{
AgentDriver object or NULL.
}
\description{
Get the Default Agent Driver
}
\examples{
\dontrun{
drv <- get_default_driver()
}
}
