% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_logic}
\alias{get_logic}
\title{Get a Registered Logic Component}
\usage{
get_logic(name)
}
\arguments{
\item{name}{String.}
}
\value{
The function or NULL if not found.
}
\description{
Get a Registered Logic Component
}
