% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_role}
\alias{get_role}
\title{Get an LLM Role (System Prompt)}
\usage{
get_role(name)
}
\arguments{
\item{name}{String. Unique identifier.}
}
\value{
String prompt text, or NULL if not found.
}
\description{
Get an LLM Role (System Prompt)
}
