% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraR.R
\docType{package}
\name{hydraR-package}
\alias{hydraR}
\alias{hydraR-package}
\title{hydraR: A framework for elegantly configuring complex applications}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Houstonwp/hydraR}
  \item Report bugs at \url{https://github.com/Houstonwp/hydraR/issues}
}

}
\author{
\strong{Maintainer}: Houston Perrett \email{hp@houstonp.com}

}
\keyword{internal}
