% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_named_list}
\alias{is_named_list}
\title{Check if an object is a named list}
\usage{
is_named_list(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
Logical.
}
\description{
Helper to verify that an object is a list (but not a data.frame) and
that all its elements have non-empty names.
}
