% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_log.R
\name{JSONLMessageLog}
\alias{JSONLMessageLog}
\title{JSONL Message Log R6 Class}
\description{
Persists messages to a JSON Lines file. Atomic file appending
ensures that multiple parallel worktree processes can log messages
without locking conflicts.
}
\section{Super class}{
\code{\link[HydraR:MessageLog]{HydraR::MessageLog}} -> \code{JSONLMessageLog}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{String. Path to JSONL file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JSONLMessageLog-new}{\code{JSONLMessageLog$new()}}
\item \href{#method-JSONLMessageLog-log}{\code{JSONLMessageLog$log()}}
\item \href{#method-JSONLMessageLog-get_all}{\code{JSONLMessageLog$get_all()}}
\item \href{#method-JSONLMessageLog-clone}{\code{JSONLMessageLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONLMessageLog-new"></a>}}
\if{latex}{\out{\hypertarget{method-JSONLMessageLog-new}{}}}
\subsection{Method \code{new()}}{
Initialize JSONLMessageLog.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONLMessageLog$new(path = tempfile(fileext = ".jsonl"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{String.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONLMessageLog-log"></a>}}
\if{latex}{\out{\hypertarget{method-JSONLMessageLog-log}{}}}
\subsection{Method \code{log()}}{
Store a message (atomic append).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONLMessageLog$log(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{List. Message object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONLMessageLog-get_all"></a>}}
\if{latex}{\out{\hypertarget{method-JSONLMessageLog-get_all}{}}}
\subsection{Method \code{get_all()}}{
Get all logs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONLMessageLog$get_all()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of logs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSONLMessageLog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JSONLMessageLog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSONLMessageLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
