% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{lint_workflow_logic}
\alias{lint_workflow_logic}
\title{Lint Workflow Logic}
\usage{
lint_workflow_logic(wf)
}
\arguments{
\item{wf}{List. The workflow object.}
}
\value{
List(errors, warnings).
}
\description{
Lint Workflow Logic
}
\keyword{internal}
