% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_log.R
\name{MemoryMessageLog}
\alias{MemoryMessageLog}
\title{Memory Message Log R6 Class}
\description{
In-memory storage for messages.
}
\section{Super class}{
\code{\link[HydraR:MessageLog]{HydraR::MessageLog}} -> \code{MemoryMessageLog}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{logs}}{List.
Log a message
Get all logs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MemoryMessageLog-log}{\code{MemoryMessageLog$log()}}
\item \href{#method-MemoryMessageLog-get_all}{\code{MemoryMessageLog$get_all()}}
\item \href{#method-MemoryMessageLog-clone}{\code{MemoryMessageLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryMessageLog-log"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryMessageLog-log}{}}}
\subsection{Method \code{log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryMessageLog$log(msg)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryMessageLog-get_all"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryMessageLog-get_all}{}}}
\subsection{Method \code{get_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryMessageLog$get_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemoryMessageLog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MemoryMessageLog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemoryMessageLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
