% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_log.R
\name{MessageLog}
\alias{MessageLog}
\title{Message Log Base R6 Class}
\value{
A \code{MessageLog} base object.
}
\description{
An abstract base class defining the interface for message logging in HydraR.
Subclasses provide concrete storage implementations (Memory, File, Database).
}
\examples{
\dontrun{
# This is an abstract base class.
# Use MemoryMessageLog or DuckDBMessageLog instead.
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MessageLog-log}{\code{MessageLog$log()}}
\item \href{#method-MessageLog-get_all}{\code{MessageLog$get_all()}}
\item \href{#method-MessageLog-clone}{\code{MessageLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageLog-log"></a>}}
\if{latex}{\out{\hypertarget{method-MessageLog-log}{}}}
\subsection{Method \code{log()}}{
Store a message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageLog$log(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{List. Message object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The log object (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageLog-get_all"></a>}}
\if{latex}{\out{\hypertarget{method-MessageLog-get_all}{}}}
\subsection{Method \code{get_all()}}{
Get all logs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageLog$get_all()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of logs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MessageLog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MessageLog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MessageLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
