% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers_cli.R
\name{OpenAICodexCLIDriver}
\alias{OpenAICodexCLIDriver}
\title{OpenAI Codex CLI Driver R6 Class}
\value{
An `OpenAICodexCLIDriver` object.
}
\description{
Driver for the official `codex` CLI tool (v0.118+).
}
\section{Super class}{
\code{\link[HydraR:AgentDriver]{HydraR::AgentDriver}} -> \code{OpenAICodexCLIDriver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OpenAICodexCLIDriver-new}{\code{OpenAICodexCLIDriver$new()}}
\item \href{#method-OpenAICodexCLIDriver-call}{\code{OpenAICodexCLIDriver$call()}}
\item \href{#method-OpenAICodexCLIDriver-clone}{\code{OpenAICodexCLIDriver$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="HydraR" data-topic="AgentDriver" data-id="exec_in_dir"><a href='../../HydraR/html/AgentDriver.html#method-AgentDriver-exec_in_dir'><code>HydraR::AgentDriver$exec_in_dir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HydraR" data-topic="AgentDriver" data-id="filter_llm_noise"><a href='../../HydraR/html/AgentDriver.html#method-AgentDriver-filter_llm_noise'><code>HydraR::AgentDriver$filter_llm_noise()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HydraR" data-topic="AgentDriver" data-id="format_cli_opts"><a href='../../HydraR/html/AgentDriver.html#method-AgentDriver-format_cli_opts'><code>HydraR::AgentDriver$format_cli_opts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HydraR" data-topic="AgentDriver" data-id="get_capabilities"><a href='../../HydraR/html/AgentDriver.html#method-AgentDriver-get_capabilities'><code>HydraR::AgentDriver$get_capabilities()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HydraR" data-topic="AgentDriver" data-id="validate_cli_opts"><a href='../../HydraR/html/AgentDriver.html#method-AgentDriver-validate_cli_opts'><code>HydraR::AgentDriver$validate_cli_opts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HydraR" data-topic="AgentDriver" data-id="validate_no_injection"><a href='../../HydraR/html/AgentDriver.html#method-AgentDriver-validate_no_injection'><code>HydraR::AgentDriver$validate_no_injection()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAICodexCLIDriver-new"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAICodexCLIDriver-new}{}}}
\subsection{Method \code{new()}}{
Initialize OpenAICodexCLIDriver
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAICodexCLIDriver$new(
  id = "codex_cli",
  validation_mode = "warning",
  working_dir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{String. Unique identifier.}

\item{\code{validation_mode}}{String. "warning" or "strict".}

\item{\code{working_dir}}{String. Optional. Path to isolated Git worktree.
Call the LLM}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAICodexCLIDriver-call"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAICodexCLIDriver-call}{}}}
\subsection{Method \code{call()}}{
Sends a execution request to the Codex CLI.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAICodexCLIDriver$call(prompt, system_prompt = NULL, cli_opts = list(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prompt}}{String.}

\item{\code{system_prompt}}{String. Optional system prompt.}

\item{\code{cli_opts}}{List.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
String. Cleaned result.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAICodexCLIDriver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAICodexCLIDriver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAICodexCLIDriver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
