% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{add_llm_node}
\alias{add_llm_node}
\title{Create an LLM Agent Node easily}
\usage{
add_llm_node(id, role, driver, ...)
}
\arguments{
\item{id}{String. Unique identifier for the node.}

\item{role}{String. System prompt/role for the agent.}

\item{driver}{AgentDriver object.}

\item{...}{Additional arguments passed to AgentLLMNode$new()}
}
\value{
AgentLLMNode object.
}
\description{
Create an LLM Agent Node easily
}
