% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{add_logic_node}
\alias{add_logic_node}
\title{Create an R Logic Node easily}
\usage{
add_logic_node(id, logic_fn, ...)
}
\arguments{
\item{id}{String. Unique identifier for the node.}

\item{logic_fn}{Function. Pure R function taking an AgentState object.}

\item{...}{Additional arguments passed to AgentLogicNode$new()}
}
\value{
AgentLogicNode object.
}
\description{
Create an R Logic Node easily
}
\examples{
\dontrun{
add_logic_node("logic1", function() print("Logic"))
}
}
