% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointer.R
\name{Checkpointer}
\alias{Checkpointer}
\title{Checkpointer Interface}
\value{
A `Checkpointer` object.
}
\description{
Abstract base class for AgentDAG checkpointers.
}
\examples{
\dontrun{
cp <- Checkpointer$new(saver = MemorySaver$new())
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Checkpointer-put}{\code{Checkpointer$put()}}
\item \href{#method-Checkpointer-get}{\code{Checkpointer$get()}}
\item \href{#method-Checkpointer-clone}{\code{Checkpointer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkpointer-put"></a>}}
\if{latex}{\out{\hypertarget{method-Checkpointer-put}{}}}
\subsection{Method \code{put()}}{
Persist state to the checkpointer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Checkpointer$put(thread_id, state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{String. Identifier for the execution thread.}

\item{\code{state}}{AgentState object. The state to save.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL (called for side effect).
Load state
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkpointer-get"></a>}}
\if{latex}{\out{\hypertarget{method-Checkpointer-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Checkpointer$get(thread_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{String. Identifier for the execution thread.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AgentState object or NULL if not found.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Checkpointer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Checkpointer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Checkpointer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
