% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{dag_add_logic_node}
\alias{dag_add_logic_node}
\title{Add an R Logic Node directly to a DAG}
\usage{
dag_add_logic_node(dag, id, logic_fn, ...)
}
\arguments{
\item{dag}{AgentDAG object.}

\item{id}{String. Unique identifier for the node.}

\item{logic_fn}{Function. Pure R function taking an AgentState object.}

\item{...}{Additional arguments passed to AgentLogicNode$new()}
}
\value{
The modified AgentDAG object (invisibly).
}
\description{
Add an R Logic Node directly to a DAG
}
\examples{
\dontrun{
dag <- dag_create()
dag <- dag_add_logic_node(dag, "node1", function() print("Hello"))
}
}
