% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{dag_create}
\alias{dag_create}
\title{Create an Agent Graph}
\usage{
dag_create(message_log = NULL)
}
\arguments{
\item{message_log}{MessageLog. Optional audit log for the DAG.}
}
\value{
An AgentDAG object.
}
\description{
Create an Agent Graph
}
