% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_log.R
\name{DuckDBMessageLog}
\alias{DuckDBMessageLog}
\title{DuckDB Message Log R6 Class}
\value{
A \code{DuckDBMessageLog} object.
}
\description{
A persistent implementation of \code{MessageLog} that writes messages to a
centralized DuckDB database. This is the recommended logger for production
and audit-heavy workflows.
}
\examples{
\dontrun{
# 1. Persistent audit logging to DuckDB
audit_log <- DuckDBMessageLog$new(
  db_path = "data/agent_audit.duckdb"
)

# 2. Attach to a DAG and run
dag <- dag_create(message_log = audit_log)
dag$run(thread_id = "experiment_404")

# 3. Retrieve and analyze messages from a previous session
prev_logs <- audit_log$get_all()
}
}
\section{Super class}{
\code{\link[HydraR:MessageLog]{HydraR::MessageLog}} -> \code{DuckDBMessageLog}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{db_path}}{String. Path to DuckDB file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DuckDBMessageLog-new}{\code{DuckDBMessageLog$new()}}
\item \href{#method-DuckDBMessageLog-log}{\code{DuckDBMessageLog$log()}}
\item \href{#method-DuckDBMessageLog-get_all}{\code{DuckDBMessageLog$get_all()}}
\item \href{#method-DuckDBMessageLog-clone}{\code{DuckDBMessageLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBMessageLog-new"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBMessageLog-new}{}}}
\subsection{Method \code{new()}}{
Finalizer to clean up the cached connection.


Initialize DuckDBMessageLog.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBMessageLog$new(db_path = "~/.gemini/memory/bot_history.duckdb")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_path}}{String.
Path to the DuckDB file. If the file does not exist, it will be
created upon the first message log. Defaults to the master
\code{bot_history.duckdb}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DuckDBMessageLog} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBMessageLog-log"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBMessageLog-log}{}}}
\subsection{Method \code{log()}}{
Store a message.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBMessageLog$log(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{List. Message object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The log object (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBMessageLog-get_all"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBMessageLog-get_all}{}}}
\subsection{Method \code{get_all()}}{
Get all logs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBMessageLog$get_all()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of logs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBMessageLog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBMessageLog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBMessageLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
