% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_log.R
\name{DuckDBMessageLog}
\alias{DuckDBMessageLog}
\title{DuckDB Message Log R6 Class}
\description{
Persists messages to the master DuckDB database.
}
\section{Super class}{
\code{\link[HydraR:MessageLog]{HydraR::MessageLog}} -> \code{DuckDBMessageLog}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{db_path}}{String. Path to DuckDB file.
Initialize DuckDBMessageLog}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DuckDBMessageLog-new}{\code{DuckDBMessageLog$new()}}
\item \href{#method-DuckDBMessageLog-log}{\code{DuckDBMessageLog$log()}}
\item \href{#method-DuckDBMessageLog-get_all}{\code{DuckDBMessageLog$get_all()}}
\item \href{#method-DuckDBMessageLog-clone}{\code{DuckDBMessageLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBMessageLog-new"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBMessageLog-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBMessageLog$new(db_path = "~/.gemini/memory/bot_history.duckdb")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_path}}{String.
Log a message
Get all logs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBMessageLog-log"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBMessageLog-log}{}}}
\subsection{Method \code{log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBMessageLog$log(msg)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBMessageLog-get_all"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBMessageLog-get_all}{}}}
\subsection{Method \code{get_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBMessageLog$get_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBMessageLog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBMessageLog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBMessageLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
