% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointer.R
\name{DuckDBSaver}
\alias{DuckDBSaver}
\title{DuckDBSaver Checkpointer}
\value{
A `DuckDBSaver` R6 object.
}
\description{
Persistent implementation of the Checkpointer interface using DuckDB.
Supports both direct DBI connections and file paths.
}
\examples{
\dontrun{
saver <- DuckDBSaver$new(db_path = "checkpoints.duckdb")
}
}
\section{Super class}{
\code{\link[HydraR:Checkpointer]{HydraR::Checkpointer}} -> \code{DuckDBSaver}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{con}}{DBIConnection.}

\item{\code{table_name}}{String.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DuckDBSaver-new}{\code{DuckDBSaver$new()}}
\item \href{#method-DuckDBSaver-put}{\code{DuckDBSaver$put()}}
\item \href{#method-DuckDBSaver-get}{\code{DuckDBSaver$get()}}
\item \href{#method-DuckDBSaver-clone}{\code{DuckDBSaver$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBSaver-new"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBSaver-new}{}}}
\subsection{Method \code{new()}}{
Initialize DuckDBSaver
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBSaver$new(con = NULL, db_path = NULL, table_name = "agent_checkpoints")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{con}}{DBIConnection. Optional if db_path is provided.}

\item{\code{db_path}}{String path to DuckDB file. Optional if con is provided.}

\item{\code{table_name}}{Name of the table to store checkpoints in.
Save state}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBSaver-put"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBSaver-put}{}}}
\subsection{Method \code{put()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBSaver$put(thread_id, state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{String.}

\item{\code{state}}{AgentState object.
Load state}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBSaver-get"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBSaver-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBSaver$get(thread_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{String.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AgentState object or NULL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DuckDBSaver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DuckDBSaver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DuckDBSaver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
