% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{get_agent_roles}
\alias{get_agent_roles}
\title{List Registered Roles}
\usage{
get_agent_roles()
}
\value{
Character vector of role names.
}
\description{
List Registered Roles
}
\examples{
\dontrun{
roles <- get_agent_roles()
}
}
