% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{list_logic}
\alias{list_logic}
\title{List Registered Components}
\usage{
list_logic()
}
\value{
Character vector of names.
}
\description{
List Registered Components
}
\examples{
\dontrun{
list_logic()
}
}
