% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpointer.R
\name{MemorySaver}
\alias{MemorySaver}
\title{MemorySaver Checkpointer}
\value{
A \code{MemorySaver} object.
}
\description{
An in-memory implementation of the \code{Checkpointer} interface. Stores
checkpoints in a dedicated R environment. This is useful for testing or
short-lived sessions where persistence to disk is not required.
}
\examples{
\dontrun{
# 1. Use MemorySaver for transient testing or short sessions
saver <- MemorySaver$new()
dag <- dag_create(checkpointer = saver)

# 2. States are saved in the 'storage' environment under thread_id
dag$run(thread_id = "test_run_01")
ls(saver$storage) # Returns "test_run_01"
}
}
\section{Super class}{
\code{\link[HydraR:Checkpointer]{HydraR::Checkpointer}} -> \code{MemorySaver}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{storage}}{Environment. Stores the states.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MemorySaver-new}{\code{MemorySaver$new()}}
\item \href{#method-MemorySaver-put}{\code{MemorySaver$put()}}
\item \href{#method-MemorySaver-get}{\code{MemorySaver$get()}}
\item \href{#method-MemorySaver-clone}{\code{MemorySaver$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemorySaver-new"></a>}}
\if{latex}{\out{\hypertarget{method-MemorySaver-new}{}}}
\subsection{Method \code{new()}}{
Initialize MemorySaver


Creates a new environment for in-memory checkpoint storage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemorySaver$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemorySaver-put"></a>}}
\if{latex}{\out{\hypertarget{method-MemorySaver-put}{}}}
\subsection{Method \code{put()}}{
Persist state to the checkpointer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemorySaver$put(thread_id, state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{String.}

\item{\code{state}}{AgentState object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemorySaver-get"></a>}}
\if{latex}{\out{\hypertarget{method-MemorySaver-get}{}}}
\subsection{Method \code{get()}}{
Load state from the checkpointer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemorySaver$get(thread_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thread_id}}{String.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
AgentState object or NULL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MemorySaver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MemorySaver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MemorySaver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
