% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag.R
\name{mermaid_to_dag}
\alias{mermaid_to_dag}
\title{Create AgentDAG from Mermaid}
\usage{
mermaid_to_dag(mermaid_str, node_factory = auto_node_factory())
}
\arguments{
\item{mermaid_str}{String. Mermaid syntax.}

\item{node_factory}{Function(id, label) -> AgentNode. Defaults to `auto_node_factory()`.}
}
\value{
The AgentDAG object.
}
\description{
Create AgentDAG from Mermaid
}
