% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers_cli.R
\name{OllamaDriver}
\alias{OllamaDriver}
\title{Ollama Driver R6 Class}
\description{
Driver for the 'ollama' CLI tool (local).
}
\section{Super class}{
\code{\link[HydraR:AgentDriver]{HydraR::AgentDriver}} -> \code{OllamaDriver}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{String. Default model.
Initialize OllamaDriver}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OllamaDriver-new}{\code{OllamaDriver$new()}}
\item \href{#method-OllamaDriver-format_cli_opts}{\code{OllamaDriver$format_cli_opts()}}
\item \href{#method-OllamaDriver-call}{\code{OllamaDriver$call()}}
\item \href{#method-OllamaDriver-clone}{\code{OllamaDriver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="HydraR" data-topic="AgentDriver" data-id="exec_in_dir"><a href='../../HydraR/html/AgentDriver.html#method-AgentDriver-exec_in_dir'><code>HydraR::AgentDriver$exec_in_dir()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HydraR" data-topic="AgentDriver" data-id="filter_llm_noise"><a href='../../HydraR/html/AgentDriver.html#method-AgentDriver-filter_llm_noise'><code>HydraR::AgentDriver$filter_llm_noise()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HydraR" data-topic="AgentDriver" data-id="get_capabilities"><a href='../../HydraR/html/AgentDriver.html#method-AgentDriver-get_capabilities'><code>HydraR::AgentDriver$get_capabilities()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HydraR" data-topic="AgentDriver" data-id="validate_cli_opts"><a href='../../HydraR/html/AgentDriver.html#method-AgentDriver-validate_cli_opts'><code>HydraR::AgentDriver$validate_cli_opts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="HydraR" data-topic="AgentDriver" data-id="validate_no_injection"><a href='../../HydraR/html/AgentDriver.html#method-AgentDriver-validate_no_injection'><code>HydraR::AgentDriver$validate_no_injection()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OllamaDriver-new"></a>}}
\if{latex}{\out{\hypertarget{method-OllamaDriver-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OllamaDriver$new(
  id = "ollama",
  model = "llama3.2",
  validation_mode = "warning",
  working_dir = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Unique identifier.}

\item{\code{model}}{String. Default model.}

\item{\code{validation_mode}}{String. "warning" or "strict".}

\item{\code{working_dir}}{String. Optional. Path to isolated Git worktree.
Format CLI Options for Ollama}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OllamaDriver-format_cli_opts"></a>}}
\if{latex}{\out{\hypertarget{method-OllamaDriver-format_cli_opts}{}}}
\subsection{Method \code{format_cli_opts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OllamaDriver$format_cli_opts(cli_opts = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cli_opts}}{List.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character vector.
Call the LLM
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OllamaDriver-call"></a>}}
\if{latex}{\out{\hypertarget{method-OllamaDriver-call}{}}}
\subsection{Method \code{call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OllamaDriver$call(prompt, model = NULL, cli_opts = list(), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prompt}}{String.}

\item{\code{model}}{String override.}

\item{\code{cli_opts}}{List.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
String. Cleaned result.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OllamaDriver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OllamaDriver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OllamaDriver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
