% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{reducer_merge_list}
\alias{reducer_merge_list}
\title{Built-in Reducer: Merge List}
\usage{
reducer_merge_list(current, new)
}
\arguments{
\item{current}{The current state list.}

\item{new}{The new list to merge.}
}
\value{
The merged list.
}
\description{
Merges two named lists using functional patterns. 
Overwrites current for matching keys.
}
