% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_registry.R
\name{set_default_driver}
\alias{set_default_driver}
\title{Set the Default Agent Driver}
\usage{
set_default_driver(driver)
}
\arguments{
\item{driver}{AgentDriver object or ID string.}
}
\value{
NULL (invisibly)
}
\description{
Set the Default Agent Driver
}
\examples{
\dontrun{
set_default_driver(AnthropicCLIDriver$new())
}
}
